Add support for linear systems stored in more partitions than the number of MPI ranks (#66)
* Add umpire memory pool options to general
* Add IJVectorReadMultipartBinary
* Add IJMatrixReadMultipartBinary
* Add HYPRE_ClearAllErrors call to LS build and preconditioner setup
* Fix dofmap array read for distributed runs
* Update ErrorMsgAdd to be variadic

+++ include/yaml.h
@@ -99,7 +99,7 @@ char* YAMLnodeFindChildValueByKey(YAMLnode*, const char*);
          { \
             _node->mapped_val = (char*) malloc(_length * sizeof(char)); \
          } \
-         else if (_length > strlen(_node->mapped_val)) \
+         else if (_length > (int) strlen(_node->mapped_val)) \
          { \
             _node->mapped_val = (char*) realloc(_node->mapped_val, _length * sizeof(char)); \
          } \
